shuffle some things around
